+++ test.js
@@ -2,7 +2,6 @@ const _ = require('lodash');
 const cssMatcher = require('jest-matcher-css');
 const postcss = require('postcss');
 const tailwindcss = require('tailwindcss');
-const defaultConfig = require('tailwindcss/defaultConfig');
 const gradientsPlugin = require('./index.js');
 
 const generatePluginCss = (config) => {
@@ -39,11 +38,19 @@ test('there is no output by default', () => {
   });
 });
 
-test('there is no output without directions', () => {
+test('there is no output without directions or positions', () => {
   return generatePluginCss({
     theme: {
-      gradients: {
+      linearGradients: {
         directions: {},
+        colors: {
+          'red': '#f00',
+          'green': '#0f0',
+          'blue': '#00f',
+        },
+      },
+      radialGradients: {
+        positions: {},
         colors: {
           'red': '#f00',
           'green': '#0f0',
@@ -59,12 +66,22 @@ test('there is no output without directions', () => {
 test('there is no output without colors', () => {
   return generatePluginCss({
     theme: {
-      gradients: {
+      linearGradients: {
         directions: {
           't': 'to top',
           'r': 'to right',
           'b': 'to bottom',
           'l': 'to left',
+        },
+        colors: {},
+      },
+      radialGradients: {
+        positions: {
+          'default': 'center',
+          't': 'top',
+          'r': 'right',
+          'b': 'bottom',
+          'l': 'left',
         },
         colors: {},
       },
@@ -74,17 +91,17 @@ test('there is no output without colors', () => {
   });
 });
 
-test('there are default directions', () => {
+test('linear gradients have default directions', () => {
   return generatePluginCss({
     theme: {
-      gradients: {
+      linearGradients: {
         colors: {
           'red': '#f00',
         },
       },
     },
     variants: {
-      gradients: [],
+      linearGradients: [],
     },
   }).then(css => {
     expect(css).toMatchCss(`
@@ -116,10 +133,56 @@ test('there are default directions', () => {
   });
 });
 
-test('directions can be customized', () => {
+test('radial gradients have default shapes, sizes, and positions', () => {
   return generatePluginCss({
     theme: {
-      gradients: {
+      radialGradients: {
+        colors: {
+          'red': '#f00',
+        },
+      },
+    },
+    variants: {
+      linearGradients: [],
+      radialGradients: [],
+    },
+  }).then(css => {
+    expect(css).toMatchCss(`
+      .bg-radial-red {
+        background-image: radial-gradient(ellipse closest-side at center, #f00, rgba(255, 0, 0, 0))
+      }
+      .bg-radial-t-red {
+        background-image: radial-gradient(ellipse closest-side at top, #f00, rgba(255, 0, 0, 0))
+      }
+      .bg-radial-tr-red {
+        background-image: radial-gradient(ellipse closest-side at top right, #f00, rgba(255, 0, 0, 0))
+      }
+      .bg-radial-r-red {
+        background-image: radial-gradient(ellipse closest-side at right, #f00, rgba(255, 0, 0, 0))
+      }
+      .bg-radial-br-red {
+        background-image: radial-gradient(ellipse closest-side at bottom right, #f00, rgba(255, 0, 0, 0))
+      }
+      .bg-radial-b-red {
+        background-image: radial-gradient(ellipse closest-side at bottom, #f00, rgba(255, 0, 0, 0))
+      }
+      .bg-radial-bl-red {
+        background-image: radial-gradient(ellipse closest-side at bottom left, #f00, rgba(255, 0, 0, 0))
+      }
+      .bg-radial-l-red {
+        background-image: radial-gradient(ellipse closest-side at left, #f00, rgba(255, 0, 0, 0))
+      }
+      .bg-radial-tl-red {
+        background-image: radial-gradient(ellipse closest-side at top left, #f00, rgba(255, 0, 0, 0))
+      }
+    `);
+  });
+});
+
+test('directions and positions can be customized', () => {
+  return generatePluginCss({
+    theme: {
+      linearGradients: {
         directions: {
           'to-top': 'to top',
         },
@@ -129,9 +192,18 @@ test('directions can be customized', () => {
           'blue': '#00f',
         },
       },
+      radialGradients: {
+        positions: {
+          'off-center': '55% 60%',
+        },
+        colors: {
+          'red': '#f00',
+        },
+      },
     },
     variants: {
-      gradients: [],
+      linearGradients: [],
+      radialGradients: [],
     },
   }).then(css => {
     expect(css).toMatchCss(`
@@ -144,6 +216,9 @@ test('directions can be customized', () => {
       .bg-gradient-to-top-blue {
         background-image: linear-gradient(to top, rgba(0, 0, 255, 0), #00f)
       }
+      .bg-radial-off-center-red {
+        background-image: radial-gradient(ellipse closest-side at 55% 60%, #f00, rgba(255, 0, 0, 0))
+      }
     `);
   });
 });
@@ -151,7 +226,7 @@ test('directions can be customized', () => {
 test('gradients can have multiple colors', () => {
   return generatePluginCss({
     theme: {
-      gradients: {
+      linearGradients: {
         directions: {
           'to-bottom': 'to bottom',
         },
@@ -160,9 +235,19 @@ test('gradients can have multiple colors', () => {
           'red-green-blue': ['#f00', '#0f0', '#00f'],
         },
       },
+      radialGradients: {
+        positions: {
+          'default': 'center',
+        },
+        colors: {
+          'red-green': ['#f00', '#0f0'],
+          'red-green-blue': ['#f00', '#0f0', '#00f'],
+        },
+      },
     },
     variants: {
-      gradients: [],
+      linearGradients: [],
+      radialGradients: [],
     },
   }).then(css => {
     expect(css).toMatchCss(`
@@ -172,14 +257,20 @@ test('gradients can have multiple colors', () => {
       .bg-gradient-to-bottom-red-green-blue {
         background-image: linear-gradient(to bottom, #f00, #0f0, #00f)
       }
+      .bg-radial-red-green {
+        background-image: radial-gradient(ellipse closest-side at center, #f00, #0f0)
+      }
+      .bg-radial-red-green-blue {
+        background-image: radial-gradient(ellipse closest-side at center, #f00, #0f0, #00f)
+      }
     `);
   });
 });
 
-test('multiple directions and multiple colors can be used together', () => {
+test('multiple directions/positions and multiple colors can be used together', () => {
   return generatePluginCss({
     theme: {
-      gradients: {
+      linearGradients: {
         directions: {
           'to-top': 'to top',
           'to-bottom': 'to bottom',
@@ -189,9 +280,20 @@ test('multiple directions and multiple colors can be used together', () => {
           'green': ['#0f0'],
         },
       },
+      radialGradients: {
+        positions: {
+          'top': 'top',
+          'bottom': 'bottom',
+        },
+        colors: {
+          'red': ['#f00'],
+          'green': ['#0f0'],
+        },
+      },
     },
     variants: {
-      gradients: [],
+      linearGradients: [],
+      radialGradients: [],
     },
   }).then(css => {
     expect(css).toMatchCss(`
@@ -207,6 +309,18 @@ test('multiple directions and multiple colors can be used together', () => {
       .bg-gradient-to-bottom-green {
         background-image: linear-gradient(to bottom, rgba(0, 255, 0, 0), #0f0)
       }
+      .bg-radial-top-red {
+        background-image: radial-gradient(ellipse closest-side at top, #f00, rgba(255, 0, 0, 0))
+      }
+      .bg-radial-bottom-red {
+        background-image: radial-gradient(ellipse closest-side at bottom, #f00, rgba(255, 0, 0, 0))
+      }
+      .bg-radial-top-green {
+        background-image: radial-gradient(ellipse closest-side at top, #0f0, rgba(0, 255, 0, 0))
+      }
+      .bg-radial-bottom-green {
+        background-image: radial-gradient(ellipse closest-side at bottom, #0f0, rgba(0, 255, 0, 0))
+      }
     `);
   });
 });
@@ -218,7 +332,7 @@ test('colors can be referenced from the theme with a closure', () => {
         'red': '#f00',
         'blue': '#00f',
       },
-      gradients: theme => ({
+      linearGradients: theme => ({
         directions: {
           'b': 'to bottom',
         },
@@ -226,7 +340,8 @@ test('colors can be referenced from the theme with a closure', () => {
       }),
     },
     variants: {
-      gradients: [],
+      linearGradients: [],
+      radialGradients: [],
     },
   }).then(css => {
     expect(css).toMatchCss(`
@@ -240,12 +355,100 @@ test('colors can be referenced from the theme with a closure', () => {
   });
 });
 
+test('radial gradient shapes and sizes can be customized', () => {
+  return generatePluginCss({
+    theme: {
+      colors: {
+        'red': '#f00',
+        'green-blue': ['#0f0', '#00f'],
+      },
+      radialGradients: theme => ({
+        shapes: {
+          'default': 'circle',
+          'ellipse': 'ellipse',
+        },
+        sizes: {
+          'default': 'closest-side',
+          'cover': 'farthest-corner',
+        },
+        positions: {
+          'default': 'center',
+          'tr': 'top right',
+        },
+        colors: theme('colors'),
+      }),
+    },
+    variants: {
+      radialGradients: [],
+    },
+  }).then(css => {
+    expect(css).toMatchCss(`
+      .bg-radial-red {
+        background-image: radial-gradient(circle closest-side at center, #f00, rgba(255, 0, 0, 0))
+      }
+      .bg-radial-ellipse-red {
+        background-image: radial-gradient(ellipse closest-side at center, #f00, rgba(255, 0, 0, 0))
+      }
+      .bg-radial-cover-red {
+        background-image: radial-gradient(circle farthest-corner at center, #f00, rgba(255, 0, 0, 0))
+      }
+      .bg-radial-ellipse-cover-red {
+        background-image: radial-gradient(ellipse farthest-corner at center, #f00, rgba(255, 0, 0, 0))
+      }
+      .bg-radial-tr-red {
+        background-image: radial-gradient(circle closest-side at top right, #f00, rgba(255, 0, 0, 0))
+      }
+      .bg-radial-ellipse-tr-red {
+        background-image: radial-gradient(ellipse closest-side at top right, #f00, rgba(255, 0, 0, 0))
+      }
+      .bg-radial-cover-tr-red {
+        background-image: radial-gradient(circle farthest-corner at top right, #f00, rgba(255, 0, 0, 0))
+      }
+      .bg-radial-ellipse-cover-tr-red {
+        background-image: radial-gradient(ellipse farthest-corner at top right, #f00, rgba(255, 0, 0, 0))
+      }
+      .bg-radial-green-blue {
+        background-image: radial-gradient(circle closest-side at center, #0f0, #00f)
+      }
+      .bg-radial-ellipse-green-blue {
+        background-image: radial-gradient(ellipse closest-side at center, #0f0, #00f)
+      }
+      .bg-radial-cover-green-blue {
+        background-image: radial-gradient(circle farthest-corner at center, #0f0, #00f)
+      }
+      .bg-radial-ellipse-cover-green-blue {
+        background-image: radial-gradient(ellipse farthest-corner at center, #0f0, #00f)
+      }
+      .bg-radial-tr-green-blue {
+        background-image: radial-gradient(circle closest-side at top right, #0f0, #00f)
+      }
+      .bg-radial-ellipse-tr-green-blue {
+        background-image: radial-gradient(ellipse closest-side at top right, #0f0, #00f)
+      }
+      .bg-radial-cover-tr-green-blue {
+        background-image: radial-gradient(circle farthest-corner at top right, #0f0, #00f)
+      }
+      .bg-radial-ellipse-cover-tr-green-blue {
+        background-image: radial-gradient(ellipse farthest-corner at top right, #0f0, #00f)
+      }
+    `);
+  });
+});
+
 test('responsive variants are generated by default', () => {
   return generatePluginCss({
     theme: {
-      gradients: {
+      linearGradients: {
         directions: {
           't': 'to top',
+        },
+        colors: {
+          'red': '#f00',
+        },
+      },
+      radialGradients: {
+        positions: {
+          'default': 'center',
         },
         colors: {
           'red': '#f00',
@@ -257,9 +460,15 @@ test('responsive variants are generated by default', () => {
       .bg-gradient-t-red {
         background-image: linear-gradient(to top, rgba(255, 0, 0, 0), #f00)
       }
+      .bg-radial-red {
+        background-image: radial-gradient(ellipse closest-side at center, #f00, rgba(255, 0, 0, 0))
+      }
       @media (min-width: 640px) {
         .sm\\:bg-gradient-t-red {
           background-image: linear-gradient(to top, rgba(255, 0, 0, 0), #f00)
+        }
+        .sm\\:bg-radial-red {
+          background-image: radial-gradient(ellipse closest-side at center, #f00, rgba(255, 0, 0, 0))
         }
       }
     `);
@@ -269,7 +478,7 @@ test('responsive variants are generated by default', () => {
 test('variants can be customized', () => {
   return generatePluginCss({
     theme: {
-      gradients: {
+      linearGradients: {
         directions: {
           't': 'to top',
         },
@@ -277,9 +486,18 @@ test('variants can be customized', () => {
           'red': '#f00',
         },
       },
+      radialGradients: {
+        positions: {
+          'b': 'bottom',
+        },
+        colors: {
+          'blue': '#00f',
+        },
+      },
     },
     variants: {
-      gradients: ['hover', 'active'],
+      linearGradients: ['hover', 'active'],
+      radialGradients: ['group-hover'],
     },
   }).then(css => {
     expect(css).toMatchCss(`
@@ -291,6 +509,12 @@ test('variants can be customized', () => {
       }
       .active\\:bg-gradient-t-red:active {
         background-image: linear-gradient(to top, rgba(255, 0, 0, 0), #f00)
+      }
+      .bg-radial-b-blue {
+        background-image: radial-gradient(ellipse closest-side at bottom, #00f, rgba(0, 0, 255, 0))
+      }
+      .group:hover .group-hover\\:bg-radial-b-blue {
+        background-image: radial-gradient(ellipse closest-side at bottom, #00f, rgba(0, 0, 255, 0))
       }
     `);
   });